Renaming: Identifiable --> Identifier Identifier   --> BasicIdentifier
Additionally:
- Better logic for deserializing JSON identifiers.
