+++ src/test/java/com/robinpowered/sdk/model/BasicIdentifierTest.java
@@ -3,11 +3,11 @@ package com.robinpowered.sdk.model;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.Test;
 
-public class IdentifierTest {
+public class BasicIdentifierTest {
 
     @Test
     public void testEqualsAndHashcode() {
-        EqualsVerifier.forClass(Identifier.class)
+        EqualsVerifier.forClass(BasicIdentifier.class)
                 .usingGetClass()
                 .verify();
     }
